Django/myproject/requirements.txt
boto==2.41.0
boto3==1.3.1
botocore==1.4.37
coreapi==1.32.0
Django==1.9.7
django-cors-headers==1.1.0
django-filter==0.13.0
django-rest-swagger==0.3.8
djangorestframework==3.4.0
docutils==0.12
futures==3.0.5
itypes==1.1.0
jmespath==0.9.0
Markdown==2.6.6
MySQL-python==1.2.5
openapi-codec==1.0.0
PyMySQL==0.7.5
python-dateutil==2.5.3
PyYAML==3.11
requests==2.10.0
simplejson==3.8.2
six==1.10.0
uritemplate==0.6
